+++ main/world_map/world_map_menu.js
@@ -619,6 +619,7 @@ function settingsMenu() {
 					menu.newEntry({
 						menuName: subMenu, entryText: 'Dynamic (background art mode)', func: () => {
 							properties.bDynamicColors[1] = !properties.bDynamicColors[1];
+							if (properties.bDynamicColors[1] && properties.bOnNotifyColors[1]) { fb.ShowPopupMessage('Warning: Dynamic colors (background art mode) and Color-server listening are enabled at the same time.\n\nThis setting may probably produce glitches since 2 color sources are being used, while one tries to override the other.\n\nIt\'s recommended to only use one of these features, unless you know what you are DOMStringList.', window.ScriptInfo.Name + ': Dynamic colors'); }
 							overwriteProperties(properties);
 							if (properties.bDynamicColors[1]) {
 								// Ensure it's applied with compatible settings
@@ -648,14 +649,19 @@ function settingsMenu() {
 							overwriteProperties(properties);
 							background.updateImageBg(true);
 							worldMap.colorsChanged();
+							if (properties.bOnNotifyColors[1]) {
+								window.NotifyOthers('Colors: ask color scheme', window.ScriptInfo.Name + ': set color scheme');
+								window.NotifyOthers('Colors: ask color', window.ScriptInfo.Name + ': set colors');
+							}
 							repaint(void (0), true);
-						}, flags: properties.bDynamicColors[1] ? MF_STRING : MF_GRAYED
+						}, flags: properties.bDynamicColors[1] || properties.bOnNotifyColors[1] ? MF_STRING : MF_GRAYED
 					});
 					menu.newCheckMenuLast(() => properties.bDynamicColorsBg[1]);
 					menu.newSeparator(subMenu);
 					menu.newEntry({
 						menuName: subMenu, entryText: 'Listen to color-servers', func: () => {
 							properties.bOnNotifyColors[1] = !properties.bOnNotifyColors[1];
+							if (properties.bDynamicColors[1] && properties.bOnNotifyColors[1]) { fb.ShowPopupMessage('Warning: Dynamic colors (background art mode) and Color-server listening are enabled at the same time.\n\nThis setting may probably produce glitches since 2 color sources are being used, while one tries to override the other.\n\nIt\'s recommended to only use one of these features, unless you know what you are DOMStringList.', window.ScriptInfo.Name + ': Dynamic colors'); }
 							overwriteProperties(properties);
 							if (properties.bOnNotifyColors[1]) {
 								window.NotifyOthers('Colors: ask color scheme', window.ScriptInfo.Name + ': set color scheme');