+++ CHANGELOG.md
@@ -65,6 +65,7 @@
 - UI: changed default background settings for a more cohesive experience along [Library-Tree-SMP](https://github.com/regorxxx/Library-Tree-SMP/), Biography and dark mode.
 - UI: improvements on dynamic colors handling in some extreme cases with main and secondary colors being almost equal.
 - UI: header text style is now italic.
+- UI: improvements to dynamic colors handling for background and server-color sources. Added warning when trying to activate both at the same time.
 ### Removed
 ### Fixed
 - UI: fixed SMP artifacts on header gradients at some sizes.